done eval

+++ verifiable_coding/verifiable_coding_880_0.py
@@ -5,7 +5,7 @@ def solve():
     input = sys.stdin.buffer.read().split()
     n = int(input[0])
     for i in range(1, n+1):
-        r = float(input[i])
+        r = int(input[i])
         area = math.pi * r * r
         print("{0:.2f}".format(area))
